pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[project]
name = "analytics-orchestrator"
version = "0.1.0"
description = "A LangGraph state machine that turns a business brief + CSV into a reproducible analytics delivery."
requires-python = ">=3.11"

dependencies = [
    # --- AI orchestration ---
    # LangGraph: the state machine — handles nodes, edges, checkpointing, human interrupts
    "langgraph>=0.2.0",
    # LangChain core: prompt templates, runnables, output parsers
    "langchain-core>=0.3.0",
    # LangChain OpenAI: ChatOpenAI with .with_structured_output() for Pydantic parsing
    "langchain-openai>=0.2.0",
    # HuggingFace provider — enables free LLM option via HuggingFace Inference API
    "langchain-huggingface>=0.1.0",

    # --- Data & ML ---
    # pandas: dataframe manipulation across ETL and profiling nodes
    "pandas>=2.2.0",
    # ydata-profiling: deterministic statistics — nulls, distributions, correlations
    "ydata-profiling>=4.9.0",
    # scikit-learn: preprocessing pipeline, models, cross-validation
    "scikit-learn>=1.5.0",
    # joblib: serialise/deserialise the trained sklearn pipeline
    "joblib>=1.4.0",

    # --- Output ---
    # streamlit: the generated dashboard artifact
    "streamlit>=1.38.0",

    # --- Infrastructure ---
    # pydantic: schema validation for every LLM output and node boundary
    "pydantic>=2.8.0",
    # python-dotenv: loads .env without hardcoding API keys
    "python-dotenv>=1.0.0",
    # rich: readable terminal output for node transitions and state summaries
    "rich>=13.8.0",
]

[project.optional-dependencies]
dev = [
    # pytest: runs golden test cases and consistency assertions
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
